Add files via upload
This Application is to register and logic to the system using Spring Security. Where user should register through the registration form and it should store the details in MYSQL database while string the details it should store the encrypted password in the database instead of plain password. once it stored the details it should be access the application by entering the Username and password fields through Login page. if the credentials matched it should give the access to the application. 
So it is complete full stack application.